Add files via upload
These apps were built with React and Mongodb since Mongodb is a cloud-based platform and there is a new transition in React.Create two folders in the directory name: frontend and backend. In the frontend folder, create files such as App.js, App.css, Login.jsx, and Register.jsx.
Create a file in the backend folder called app.js or UserDetails.js.